Restructure files
See also: #28
