+++ wlm-ui/src/pages/MainPage/MainPage.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { useDispatch } from 'react-redux';
 
-import { AppDispatch } from '../store';
-import { signout } from '../store/slices/user/user';
-import { calibrate } from '../store/slices/calibration/calibration';
-import ChannelList from './Channel/ChannelList';
-import EventList from './EventList/EventList';
+import { AppDispatch } from '../../store';
+import { signout } from '../../store/slices/user/user';
+import { calibrate } from '../../store/slices/calibration/calibration';
+import ChannelList from './components/ChannelList/ChannelList';
+import EventList from './components/EventList/EventList';
 
 const MainPage = () => {
     const dispatch = useDispatch<AppDispatch>();
