+++ wlm-ui/src/pages/MainPage/components/EventList/EventList.scss
